[NA] [BE] Refactor Evolutionary Optimizer in Opik Optimizer SDK (#3202)
* Create AGENTS.md

* refactor: evolutionary optimizer

* refactor: evolutionary optimizer

* refactor: file name

* chore: warning for py3.13

* fix: missing failed prompt for evolutionary optimizer

* fix: better error handling of Litellm on evolutionary optimizer

* chore: lint

* chore: JSON extraction

* fix: types for evolutionary optimizer

* Update sdks/opik_optimizer/src/opik_optimizer/evolutionary_optimizer/mutation_ops.py


* Update sdks/opik_optimizer/src/opik_optimizer/evolutionary_optimizer/llm_support.py


* Update sdks/opik_optimizer/src/opik_optimizer/evolutionary_optimizer/mutation_ops.py


* Update sdks/opik_optimizer/src/opik_optimizer/evolutionary_optimizer/crossover_ops.py


* chore: revert bad change from copilot

* Update evolutionary_optimizer.py

* Create AGENTS.md

* refactor: evolutionary optimizer

* refactor: evolutionary optimizer

* refactor: file name

* chore: warning for py3.13

* fix: missing failed prompt for evolutionary optimizer

* fix: better error handling of Litellm on evolutionary optimizer

* chore: lint

* chore: JSON extraction

* fix: types for evolutionary optimizer

* Update sdks/opik_optimizer/src/opik_optimizer/evolutionary_optimizer/mutation_ops.py


* Update sdks/opik_optimizer/src/opik_optimizer/evolutionary_optimizer/llm_support.py


* Update sdks/opik_optimizer/src/opik_optimizer/evolutionary_optimizer/mutation_ops.py


* Update sdks/opik_optimizer/src/opik_optimizer/evolutionary_optimizer/crossover_ops.py


* chore: revert bad change from copilot

* Update evolutionary_optimizer.py

* Delete sdks/opik_optimizer/AGENTS.md

---------
